Close Consumer tasks explicitly
Cancels all consumer tasks that might be active when `finalize` method
is called. If they are not cancelled `AioKafka` does not cancel neither
actively the connection to kafka.

During testing without having this addressed can cause long periods of
rebalancing between test executions since Kafka/Zookeper does not know
what is the status of the last consumer used in the last test.